pyproject.toml
[project]
name = "crewai-news-bias-agent-workflow"
version = "0.1.0"
description = "A CrewAI-powered orchestration that connects to news article input, analyzes news and bias data and prepares basic scoring. Built with modular CrewAI agent patterns."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"

dependencies = [
    "chromadb~=0.6.0",
    "crewai>=0.114.0",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pytz~=2025.1",
    "requests>=2.32.3",
    "uv>=0.6.14",
    "pyyaml>=6.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["trust_and_bias_analysis"]

[tool.setuptools.package-data]
"trust_and_bias_analysis.config" = ["*.yaml"]